Refactoring View Models to leverage MVVM Toolkit

+++ App/ViewModels/DiagnoseViewModel.cs
@@ -1,101 +1,61 @@
 ﻿using System.Collections.ObjectModel;
-using System.ComponentModel.DataAnnotations;
 using System.Windows.Input;
 using App.Contracts.Services;
 using App.Contracts.ViewModels;
 using App.Core.Models;
-using App.Core.Services;
 using App.Core.Services.Interfaces;
-using App.Services;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 
 namespace App.ViewModels;
 
-public class DiagnoseViewModel : ObservableObject, INavigationAware
+public partial class DiagnoseViewModel : ObservableObject, INavigationAware
 {
-
+    [ObservableProperty]
     private Diagnose _selectedItem;
-    public Diagnose SelectedItem
-    {
-        get => _selectedItem;
-        set
-        {
-            if (value != null)
-            {
-                _selectedItem = value;
 
-            }
-        }
-    }
-
-
-
-
+    [ObservableProperty]
     private ObservableCollection<Diagnose> _diagnoses = new();
-    public ObservableCollection<Diagnose> Diagnoses
-    {
-        get => _diagnoses;
-        set => _diagnoses = value;
-    }
 
     private readonly ICrudService<Diagnose> _crudService;
-
     private readonly INavigationService _navigationService;
-    public IInfoBarService InfoBarService
-    {
-        get;
-    }
-
-    public ICommand DeleteDiagnoseCommand
-    {
-        get;
-    }
+    private readonly IDialogService _dialogService;
+    public IInfoBarService _infoBarService;
 
 
-    public ICommand NavigateToUpdateDiagnoseCommand
-    {
-        get;
-    }
-
-
-
-    public DiagnoseViewModel(ICrudService<Diagnose> crudService, IInfoBarService infoBarService, INavigationService navigationService)
+    public DiagnoseViewModel(ICrudService<Diagnose> crudService, IInfoBarService infoBarService, INavigationService navigationService, IDialogService dialogService)
     {
         _crudService = crudService;
-        InfoBarService = infoBarService;
+        _infoBarService = infoBarService;
         _navigationService = navigationService;
-        DeleteDiagnoseCommand = new RelayCommand(DeleteDiagnose);
-        NavigateToUpdateDiagnoseCommand = new RelayCommand<Diagnose>(NavigateToUpdateDiagnose);
-        Diagnoses = new ObservableCollection<Diagnose>();
+        _dialogService = dialogService;
 
     }
-
-    private async void DeleteDiagnose()
+    [RelayCommand]
+    public async void Delete()
     {
-        Diagnose diagnoseToRemove = SelectedItem;
-        Diagnoses.Remove(diagnoseToRemove);
-        await _crudService.Delete(diagnoseToRemove);
-        InfoBarService.showMessage("Erfolgreich gelöscht", "Erfolgreich");
+        var result = await _dialogService.ConfirmDeleteDialogAsync("Fehlerkategorie löschen", "Sind sie sicher, dass sie diesen Eintrag löschen wollen?");
+        if (result != null && result == true)
+        {
+            Diagnose diagnoseToRemove = SelectedItem;
+            Diagnoses.Remove(diagnoseToRemove);
+            await _crudService.Delete(diagnoseToRemove);
+            _infoBarService.showMessage("Erfolgreich gelöscht", "Erfolgreich");
+        }
 
     }
 
-    private async void NavigateToUpdateDiagnose(Diagnose diagnose)
+    [RelayCommand]
+    public void NavigateToUpdate(Diagnose diagnose)
     {
         _navigationService.NavigateTo("App.ViewModels.UpdateDiagnoseViewModel", diagnose);
-        /*var item = Diagnoses.FirstOrDefault(i => i == _selectedItem);
-        if (item != null)
-        {
-            item = _selectedItem;
-        }*/
 
     }
 
     public async void OnNavigatedTo(object parameter)
     {
-        Diagnoses.Clear();
+        _diagnoses.Clear();
 
-        // TODO: Replace with real data.
         var response = await _crudService.GetAll();
 
 
@@ -103,12 +63,12 @@ public class DiagnoseViewModel : ObservableObject, INavigationAware
         {
             foreach (var item in response.Data)
             {
-                Diagnoses.Add(item);
+                _diagnoses.Add(item);
             }
         }
         else
         {
-            InfoBarService.showError("ErrorMessage", response.ErrorMessage);
+            _infoBarService.showError("Daten konnten nicht geladen werden", "Error");
         }
     }
